[karabiner] use left cmd as hyper
I was using "j" and "k" simultaneously to trigger the hyperkey. It
wasn't working for me because I kept triggering it by mistake. I'm now
replacing it with command-left, but only when pressed alone. I already
noticed a drawback, which that sometimes I cannot command click to open
links in new tabs.